+++ clients/test/unitcore/test_unit_node_based_access.py
@@ -281,3 +281,17 @@ class test_node_based_access(unittest.TestCase):
 
         expected_result = "underscores help text"
         self.assertEqual(expected_result, yangvoodoo.DataAccess.get_extension(self.root.underscoretests, 'info', 'underscore_only'))
+
+    def test_choices(self):
+        self.assertEqual(repr(self.root.morecomplex.inner.beer_type), "VoodooChoice{/integrationtest:morecomplex/inner/...beer_type}")
+        self.assertEqual(repr(self.root.morecomplex.inner.beer_type.craft), "VoodooCase{/integrationtest:morecomplex/inner/...craft}")
+
+        self.root.morecomplex.inner.beer_type.craft.brewdog = "PUNK IPA"
+        self.assertEqual(self.root.morecomplex.inner.beer_type.craft.brewdog, "PUNK IPA")
+        self.assertEqual(self.stub.stub_store['/integrationtest:morecomplex/inner/brewdog'], 'PUNK IPA')
+
+    def test_silly_things(self):
+        with self.assertRaises(yangvoodoo.Errors.CannotAssignValueToContainingNode) as context:
+            self.root.morecomplex = 'ssdfsdf'
+        self.assertEqual(str(context.exception), "Cannot assign a value to morecomplex")
+        self.assertNotEqual(self.root.morecomplex.inner.beer_type.craft.brewdog, "PUNK IPA")