+++ clients/yangvoodoo/basedal.py
@@ -202,3 +202,6 @@ class BaseDataAbstractionLayer:
 
     def has_datastore_changed(self):
         raise NotImplementedError("has_datastore_changed not implemented")
+
+    def dump_xpaths(self):
+        raise NotImplementedError("dump_xpaths not implemented")